IndexExistsCheck and removed package steps
