+++ charles-rest/src/main/java/com/amihaiemil/charles/github/AuthorOwnerCheck.java
@@ -26,10 +26,8 @@
 package com.amihaiemil.charles.github;
 
 import java.io.IOException;
-import org.slf4j.Logger;
 
-import com.amihaiemil.charles.steps.PreconditionCheckStep;
-import com.amihaiemil.charles.steps.Step;
+import org.slf4j.Logger;
 
 /**
  * Step where it's checked if the repo is under the author's name.
