+++ charles-rest/src/main/java/com/amihaiemil/charles/github/IndexPage.java
@@ -22,31 +22,27 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-
-package com.amihaiemil.charles.steps;
+package com.amihaiemil.charles.github;
 
 import java.io.IOException;
+import java.util.Arrays;
+
 import org.slf4j.Logger;
+
 import com.amihaiemil.charles.DataExportException;
-import com.amihaiemil.charles.GraphCrawl;
-import com.amihaiemil.charles.IgnoredPatterns;
-import com.amihaiemil.charles.WebCrawl;
+import com.amihaiemil.charles.LiveWebPage;
+import com.amihaiemil.charles.SnapshotWebPage;
+import com.amihaiemil.charles.WebPage;
 import com.amihaiemil.charles.aws.AmazonEsRepository;
-import com.amihaiemil.charles.github.Command;
 
 /**
- * Step to index a website.
+ * Step to index a single page.
  * @author Mihai Andronache (amihaiemil@gmail.com)
  * @version $Id$
  * @since 1.0.0
  *
  */
-public class IndexSite extends IndexStep {
-
-    /**
-     * Command.
-     */
-    private Command com;
+public class IndexPage extends IndexStep {
 
     /**
      * Action's logger.
@@ -54,12 +50,17 @@ public class IndexSite extends IndexStep {
     private Logger logger;
 
     /**
-     * Constructor.
-     * @param com Command
-     * @param logger The action's logger
-     * @param next The next step to take
+     * Command.
      */
-    public IndexSite(Command com, Logger logger, Step next) {
+    private Command com;
+
+    /**
+     * Ctor.
+     * @param com Given command.
+     * @param logger Logger.
+     * @param next Next step to take.
+     */
+    public IndexPage(Command com, Logger logger, Step next) {
         super(next);
         this.com = com;
         this.logger = logger;
@@ -67,36 +68,31 @@ public class IndexSite extends IndexStep {
 
     @Override
     public void perform() {
-        try {
-        	logger.info("Starting to index the whole site...");
-            this.graphCrawl().crawl();
-        	logger.info("Indexing finished successfully!");
-        } catch (
-            DataExportException |
-            IOException |
-            RuntimeException e 
-        ) {
-            logger.error("Exception while indexing the website!", e);
-            throw new IllegalStateException("Exception while indexing the website", e);
+	    String link = this.getLink();
+	    logger.info("Indexing page " + link + " ...");
+    	try {
+    	    logger.info("Crawling the page...");
+            WebPage snapshot = new SnapshotWebPage(
+                new LiveWebPage(this.phantomJsDriver(), link)
+            );
+    	    logger.info("Page crawled. Sending to aws...");
+            new AmazonEsRepository(this.com.indexName()).export(Arrays.asList(snapshot));
+    	    logger.info("Page successfully sent to aws!");
+        } catch (DataExportException | IOException | RuntimeException e) {
+            logger.error("Exception while indexing the page " + link, e);
+            throw new IllegalStateException("Exception while indexing the page" + link, e);
         }
         this.next().perform();
     }
 
-    public WebCrawl graphCrawl() throws IOException {
-        String repoName = this.com.repo().name();
-        String siteIndexUrl;
-        if(com.repo().hasGhPagesBranch()) {
-            siteIndexUrl = "http://" + com.authorLogin() + ".github.io/" + repoName;
-        } else {
-            siteIndexUrl = "http://" + repoName;
-        }
-        logger.info("Graph-crawling, starting from " + siteIndexUrl
-        + " .The website will be crawled as a graph, going in-depth from the index page.");
-        WebCrawl siteCrawl = new GraphCrawl(
-            siteIndexUrl, this.phantomJsDriver(), new IgnoredPatterns(),
-            new AmazonEsRepository(this.com.indexName()), 20
-        );
-        return siteCrawl;
+    /**
+     * Get the page's link from the command's text which should be in markdown format, with a
+     * link like [this](http://link.com/here/the/ling) .
+     * @return String link.
+     */
+    private String getLink() {
+    	String body = this.com.json().getString("body");
+        return body.substring(body.indexOf('('),  body.indexOf(')'));
     }
 
 }
