+++ charles-rest/src/main/java/com/amihaiemil/charles/github/SendReply.java
@@ -29,9 +29,6 @@ import java.io.IOException;
 
 import org.slf4j.Logger;
 
-import com.amihaiemil.charles.steps.IntermediaryStep;
-import com.amihaiemil.charles.steps.Step;
-
 /**
  * Step where the agent sends a text reply.
  * @author Mihai Andronache (amihaiemil@gmail.com)
