+++ charles-rest/src/main/resources/responses_en.properties
@@ -24,6 +24,11 @@ denied.deleteindex.comment=@%s the repository's name in a delete command has to 
                            and it has to match the actual reponame exactly (case sensitive) \n\n\
                            A valid delete command here is ``@%s delete `%s` index``.
 
+index.missing.comment=@%s there is no index for this repo, so your command\n\
+                      cannot be fulfilled. Either the index was removed already\n\
+                      following a ``delete`` command, or it has never existed.\n\n\
+                      Check the [logs](%s) for more details.
+
 step.failure.comment=@%s Some steps failed when precessing your command. See [logs](%s) for details.
 
 index.finished.comment=@%s Index command executed successfully! \n \
