+++ charles-rest/src/test/java/com/amihaiemil/charles/github/AuthorOwnerCheckTestCase.java
@@ -34,8 +34,6 @@ import org.junit.Test;
 import org.mockito.Mockito;
 import org.slf4j.Logger;
 
-import com.amihaiemil.charles.steps.Step;
-
 /**
  * Unit tests for {@link AuthorOwnerCheck}
  * @author Mihai Andronache (amihaiemil@gmail.com)
