+++ charles-rest/src/test/java/com/amihaiemil/charles/github/DeleteIndexCommandCheckTestCase.java
@@ -25,10 +25,10 @@
 package com.amihaiemil.charles.github;
 
 import javax.json.Json;
+
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.slf4j.Logger;
-import com.amihaiemil.charles.steps.Step;
 
 /**
  * Unit tests for {@link DeleteIndexCommandCheck}
