+++ charles-rest/src/test/java/com/amihaiemil/charles/github/GhPagesBranchCheckTestCase.java
@@ -31,8 +31,6 @@ import org.junit.Test;
 import org.mockito.Mockito;
 import org.slf4j.Logger;
 
-import com.amihaiemil.charles.steps.Step;
-
 /**
  * Unit tests for {@link GhPagesBranchCheck}
  * @author Mihai Andronache (amihaiemil@gmail.com)
