+++ charles-rest/src/test/java/com/amihaiemil/charles/github/IndexPageaActionTestCase.java
@@ -37,8 +37,6 @@ import org.junit.Test;
 import org.mockito.Mockito;
 import org.slf4j.Logger;
 
-import com.amihaiemil.charles.steps.IndexPage;
-import com.amihaiemil.charles.steps.Step;
 import com.google.common.collect.Lists;
 import com.jcabi.github.Comment;
 import com.jcabi.github.Coordinates;
