+++ charles-rest/src/test/java/com/amihaiemil/charles/github/PageHostedOnGithubCheckTestCase.java
@@ -25,12 +25,13 @@
 package com.amihaiemil.charles.github;
 
 import java.io.IOException;
+
 import javax.json.Json;
 import javax.json.JsonObject;
+
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.slf4j.Logger;
-import com.amihaiemil.charles.steps.Step;
 
 /**
  * Unit tests for {@link PageHostedOnGithubCheck}
